+++ src/settings/ManageAuthoritySourceFiles/getFieldComponents.js
@@ -9,45 +9,54 @@ import { authorityFilesColumns } from './constants';
 
 /* eslint-disable react/prop-types */
 /* eslint-disable react/no-multi-comp */
-export const getFieldComponents = (selectableFieldLabel, intl) => ({
-  [authorityFilesColumns.NAME]: ({ fieldProps }) => (
+export const getFieldComponents = fieldLabels => ({
+  [authorityFilesColumns.NAME]: ({ fieldProps, name, rowIndex, fieldIndex, error }) => (
     <Field
-      {...fieldProps}
+      name={fieldProps.name}
       component={TextField}
-      placeholder={intl.formatMessage({ id: 'ui-marc-authorities.settings.manageAuthoritySourceFiles.column.name' })}
+      aria-label={`${fieldLabels[name]} ${rowIndex}`}
+      placeholder={fieldLabels[name]}
       marginBottom0
+      autoFocus={fieldIndex === 0}
+      error={error?.fieldErrors?.[name]}
     />
   ),
-  [authorityFilesColumns.CODES]: ({ fieldProps }) => (
+  [authorityFilesColumns.CODES]: ({ fieldProps, name, rowIndex, error }) => (
     <Field
-      {...fieldProps}
+      name={fieldProps.name}
+      aria-label={`${fieldLabels[name]} ${rowIndex}`}
       component={TextField}
-      placeholder={intl.formatMessage({ id: 'ui-marc-authorities.settings.manageAuthoritySourceFiles.column.codes' })}
+      placeholder={fieldLabels[name]}
       marginBottom0
+      error={error?.fieldErrors?.[name]}
     />
   ),
-  [authorityFilesColumns.START_NUMBER]: ({ fieldProps }) => (
+  [authorityFilesColumns.START_NUMBER]: ({ fieldProps, name, rowIndex, error }) => (
     <Field
-      {...fieldProps}
+      name={fieldProps.name}
+      aria-label={`${fieldLabels[name]} ${rowIndex}`}
       component={TextField}
-      placeholder={intl.formatMessage({ id: 'ui-marc-authorities.settings.manageAuthoritySourceFiles.column.startNumber' })}
+      placeholder={fieldLabels[name]}
       marginBottom0
+      error={error?.fieldErrors?.[name]}
     />
   ),
-  [authorityFilesColumns.BASE_URL]: ({ fieldProps }) => (
+  [authorityFilesColumns.BASE_URL]: ({ fieldProps, name, rowIndex, error }) => (
     <Field
-      {...fieldProps}
+      name={fieldProps.name}
+      aria-label={`${fieldLabels[name]} ${rowIndex}`}
       component={TextField}
-      placeholder={intl.formatMessage({ id: 'ui-marc-authorities.settings.manageAuthoritySourceFiles.column.baseUrl' })}
+      placeholder={fieldLabels[name]}
       marginBottom0
+      error={error?.fieldErrors?.[name]}
     />
   ),
-  [authorityFilesColumns.SELECTABLE]: ({ fieldProps }) => (
+  [authorityFilesColumns.SELECTABLE]: ({ fieldProps, name, rowIndex }) => (
     <Field
-      {...fieldProps}
+      name={fieldProps.name}
+      type="checkbox"
+      aria-label={`${fieldLabels[name]} ${rowIndex}`}
       component={Checkbox}
-      aria-label={selectableFieldLabel}
-      marginBottom0
     />
   ),
 });