+++ test/jest/__mock__/stripesCore.mock.js
@@ -105,6 +105,8 @@ jest.mock('@folio/stripes/core', () => {
 
   const AppContextMenu = ({ children }) => <>{children()}</>;
 
+  const checkIfUserInMemberTenant = jest.fn(jest.requireActual('@folio/stripes/core').checkIfUserInMemberTenant);
+
   STRIPES.connect = stripesConnect;
 
   return {
@@ -116,6 +118,7 @@ jest.mock('@folio/stripes/core', () => {
     useOkapiKy,
     useNamespace,
     useStripes: jest.fn().mockReturnValue(STRIPES),
+    checkIfUserInMemberTenant,
   };
 }, { virtual: true });